major fix to fifo, groupnorm2d now passing
