+++ machop/mase_components/norm/test/group_norm_2d_tb.py
@@ -218,7 +218,7 @@ class ErrorThresholdStreamMonitor(StreamMonitor):
                 assert fail, "Test Failed!"
                 return
 
-        self.log.info(
+        self.log.debug(
             "Passed | Got: %20s Exp: %20s Err: %10s" % (g, e, err)
         )
 
@@ -274,42 +274,6 @@ class GroupNorm2dTB(Testbench):
             check=False,
         )
 
-        # Intermediate value monitors
-        # self.mu_monitor = StreamMonitor(
-        #     dut.clk, dut.mu_in, dut.mu_acc_valid, dut.mu_acc_ready,
-        #     name="Mu Monitor",
-        # )
-        # self.squares_monitor = StreamMonitor(
-        #     dut.clk, dut.square_out, dut.squares_adder_tree_in_valid, dut.squares_adder_tree_in_ready,
-        #     name="Squares Monitor",
-        # )
-        # self.sum_squares_monitor = StreamMonitor(
-        #     dut.clk, dut.squares_acc, dut.squares_acc_valid, dut.squares_acc_ready,
-        #     name="Sum Squares Monitor",
-        # )
-        # self.var_monitor = StreamMonitor(
-        #     dut.clk, dut.variance_in, dut.squares_acc_valid, dut.squares_acc_ready,
-        #     name="Variance Monitor",
-        # )
-        # self.var_clamp_monitor = StreamMonitor(
-        #     dut.clk, dut.variance_clamp_out, dut.variance_clamp_valid,
-        #     dut.variance_clamp_ready, name="Variance Clamp Monitor",
-        # )
-        # self.isqrt_monitor = StreamMonitor(
-        #     dut.clk, dut.inv_sqrt_data, dut.inv_sqrt_valid, dut.inv_sqrt_ready,
-        #     name="Inverse Sqrt Monitor",
-        # )
-        # self.diff_monitor = StreamMonitor(
-        #     dut.clk, dut.diff_batch_out, dut.diff_batch_out_valid, dut.diff_batch_out_ready,
-        #     name="Diff Monitor",
-        # )
-        # self.norm_monitor = StreamMonitor(
-        #     dut.clk, dut.norm_out_data, dut.norm_monitor_valid, dut.norm_monitor_ready,
-        #     name="Norm Monitor",
-        # )
-
-
-
     def generate_inputs(self, batches=1):
         inputs = list()
         for _ in range(self.total_channels * batches):
@@ -318,35 +282,8 @@ class GroupNorm2dTB(Testbench):
             ))
         return inputs
 
-    # def load_intermediate(self, intermediate):
-    #     self.mu_monitor.load_monitor(intermediate["mu"])
-    #     self.var_monitor.load_monitor(intermediate["var"])
-    #     self.var_clamp_monitor.load_monitor(intermediate["var_clamp"])
-    #     self.isqrt_monitor.load_monitor(intermediate["isqrt"])
-    #     self.sum_squares_monitor.load_monitor(intermediate["sum_squares"])
-    #     diff = intermediate["diff"].reshape(-1, self.TOTAL_DIM1, self.TOTAL_DIM0)
-    #     squares = intermediate["squares"].reshape(-1, self.TOTAL_DIM1, self.TOTAL_DIM0)
-    #     norm = intermediate["norm"].reshape(-1, self.TOTAL_DIM1, self.TOTAL_DIM0)
-
-    #     def _load(mon, tensor):
-    #         for i in range(tensor.shape[0]):
-    #             mon.load_monitor(
-    #                 split_matrix(tensor[i], *self.total_tup, *self.compute_tup)
-    #             )
-    #     _load(self.diff_monitor, diff)
-    #     _load(self.squares_monitor, squares)
-    #     _load(self.norm_monitor, norm)
-
     def assert_all_monitors_empty(self):
         assert self.output_monitor.exp_queue.empty()
-        # assert self.mu_monitor.exp_queue.empty()
-        # assert self.var_monitor.exp_queue.empty()
-        # assert self.var_clamp_monitor.exp_queue.empty()
-        # assert self.isqrt_monitor.exp_queue.empty()
-        # assert self.diff_monitor.exp_queue.empty()
-        # assert self.norm_monitor.exp_queue.empty()
-        # assert self.squares_monitor.exp_queue.empty()
-        # assert self.sum_squares_monitor.exp_queue.empty()
 
 
     def model(self, inputs):
@@ -361,24 +298,6 @@ class GroupNorm2dTB(Testbench):
 
         # Float Model
         float_y = self.quantized_model(x)
-        # norm_out_float, norm_int_out, intermediate = _fixed_group_norm_2d_model(
-        #     x=x,
-        #     in_width=self.IN_WIDTH,
-        #     in_frac_width=self.IN_FRAC_WIDTH,
-        #     diff_width=self.DIFF_WIDTH,
-        #     diff_frac_width=self.DIFF_FRAC_WIDTH,
-        #     square_width=self.SQUARE_WIDTH,
-        #     square_frac_width=self.SQUARE_FRAC_WIDTH,
-        #     variance_width=self.VARIANCE_WIDTH,
-        #     variance_frac_width=self.VARIANCE_FRAC_WIDTH,
-        #     isqrt_width=self.ISQRT_WIDTH,
-        #     isqrt_frac_width=self.ISQRT_FRAC_WIDTH,
-        #     isqrt_lut=self.isqrt_lut,
-        #     norm_width=self.NORM_WIDTH,
-        #     norm_frac_width=self.NORM_FRAC_WIDTH,
-        #     out_width=self.OUT_WIDTH,
-        #     out_frac_width=self.OUT_FRAC_WIDTH,
-        # )
 
         # Output beat reconstruction
         y = integer_quantizer_for_hw(float_y, self.OUT_WIDTH, self.OUT_FRAC_WIDTH)
@@ -394,7 +313,6 @@ class GroupNorm2dTB(Testbench):
         self.in_driver.load_driver(inputs)
         exp_out = self.model(inputs)
         self.output_monitor.load_monitor(exp_out)
-        # self.load_intermediate(intermediate)
 
 
 @cocotb.test()
@@ -416,84 +334,41 @@ async def stream(dut):
     await Timer(200, 'us')
     assert tb.output_monitor.exp_queue.empty()
 
-# @cocotb.test(skip=True)
-# async def long_stream(dut):
-#     tb = GroupNorm2dTB(dut)
-#     await tb.reset()
-#     tb.output_monitor.ready.value = 1
 
-#     inputs = tb.generate_inputs(num=1000)
-#     tb.in_driver.load_driver(inputs)
-#     exp_out = tb.model(inputs)
-#     tb.output_monitor.load_monitor(exp_out)
-
-#     await Timer(2000, 'us')
-#     assert tb.output_monitor.exp_queue.empty()
+@cocotb.test()
+async def backpressure(dut):
+    tb = GroupNorm2dTB(dut)
+    await tb.reset()
+    cocotb.start_soon(bit_driver(dut.out_ready, dut.clk, 0.5))
+    tb.setup_test(batches=200)
+    await Timer(400, 'us')
+    assert tb.output_monitor.exp_queue.empty()
 
 
-# @cocotb.test(skip=True)
-# async def backpressure(dut):
-#     tb = GroupNorm2dTB(dut)
-#     await tb.reset()
-#     cocotb.start_soon(bit_driver(dut.out_ready, dut.clk, 0.5))
-
-#     inputs = tb.generate_inputs(num=200)
-#     tb.in_driver.load_driver(inputs)
-#     exp_out = tb.model(inputs)
-#     tb.output_monitor.load_monitor(exp_out)
-
-#     await Timer(400, 'us')
-#     assert tb.output_monitor.exp_queue.empty()
+@cocotb.test()
+async def valid_toggle(dut):
+    tb = GroupNorm2dTB(dut)
+    await tb.reset()
+    tb.output_monitor.ready.value = 1
+    tb.in_driver.set_valid_prob(0.5)
+    tb.setup_test(batches=200)
+    await Timer(400, 'us')
+    assert tb.output_monitor.exp_queue.empty()
 
 
-# @cocotb.test(skip=True)
-# async def valid_toggle(dut):
-#     tb = GroupNorm2dTB(dut)
-#     await tb.reset()
-#     tb.output_monitor.ready.value = 1
-#     tb.in_driver.set_valid_prob(0.5)
+@cocotb.test()
+async def valid_backpressure(dut):
+    tb = GroupNorm2dTB(dut)
+    await tb.reset()
+    tb.in_driver.set_valid_prob(0.5)
+    cocotb.start_soon(bit_driver(dut.out_ready, dut.clk, 0.5))
+    tb.setup_test(batches=200)
 
-#     inputs = tb.generate_inputs(num=200)
-#     tb.in_driver.load_driver(inputs)
-#     exp_out = tb.model(inputs)
-#     tb.output_monitor.load_monitor(exp_out)
-
-#     await Timer(400, 'us')
-#     assert tb.output_monitor.exp_queue.empty()
-
-
-# @cocotb.test(skip=True)
-# async def valid_backpressure(dut):
-#     tb = GroupNorm2dTB(dut)
-#     await tb.reset()
-#     tb.in_driver.set_valid_prob(0.5)
-#     cocotb.start_soon(bit_driver(dut.out_ready, dut.clk, 0.5))
-
-#     inputs = tb.generate_inputs(num=200)
-#     tb.in_driver.load_driver(inputs)
-#     exp_out = tb.model(inputs)
-#     tb.output_monitor.load_monitor(exp_out)
-
-#     await Timer(400, 'us')
-#     assert tb.output_monitor.exp_queue.empty()
+    await Timer(400, 'us')
+    assert tb.output_monitor.exp_queue.empty()
 
 
 if __name__ == "__main__":
-    # def calculate_isqrt_width(
-    #     total_dim0, total_dim1, compute_dim0, compute_dim1, group_channels,
-    #     in_width, **kwargs
-    # ):
-    #     depth_dim0 = total_dim0 // compute_dim0
-    #     depth_dim1 = total_dim1 // compute_dim1
-    #     square_width = in_width * 2
-    #     squares_adder_tree_in_size = compute_dim0 * compute_dim1
-    #     squares_adder_tree_out_width = ceil(log2(squares_adder_tree_in_size)) + square_width
-    #     num_iters = depth_dim0 * depth_dim1 * group_channels
-    #     iter_width = ceil(log2(num_iters))
-    #     isqrt_width = iter_width + squares_adder_tree_out_width
-    #     return isqrt_width
-
-
     def gen_cfg():
         # compute_dim0, compute_dim1, total_dim0, total_dim1 = random_2d_dimensions()
         params = {
@@ -501,14 +376,12 @@ if __name__ == "__main__":
             "TOTAL_DIM1": 4,
             "COMPUTE_DIM0": 2,
             "COMPUTE_DIM1": 2,
-            "GROUP_CHANNELS": 2, # randint(1, 4),
+            "GROUP_CHANNELS": 2,
             "IN_WIDTH": 8,
             "IN_FRAC_WIDTH": 4,
             "OUT_WIDTH": 8,
             "OUT_FRAC_WIDTH": 4,
         }
-        # lower_case_keys = {k.lower(): v for k, v in params.items()}
-        # lut_width = calculate_isqrt_width(**lower_case_keys)
         params |= lut_parameter_dict(2**5, 16)
         return params
